# Pre-merge label,jump,label,jump,... instructions