This optimizes the fixpoint calculation which
otherwise would require N*N passes in order to
optimize N labels from patterns.

I have used [this code](https://gist.github.com/josevalim/399302f7de8849b44aa7), extracted from a more complex example,
to pin down the slow down we were seeing during compilation.
The fprof results confirms the N*N behaviour given set operations
are invoked 1_000_000 times for 1_000 clauses.

After applying this optimization, compilation of the slow.erl file in the
gist above went from 0.5s to 0.02s. However, a file with 10_000 clauses
went from 60s to 0.5s.