@bjorng 2015-02-17
Only looking at the commit message and the code, it is not at all obvious what kind of code would cause the O(N*N) run-time. So I would like a short code example directly in the commit message. An extract from the gist would do.

Also, a few sentences about how much compilation time this optimisation saved directly in the commit message would be good. Basically, the information already present in the pull request.