@bjorng 2015-02-18
Thanks! We will test it a few days in our daily builds and then merge it to master.