@nox 2015-02-18
Can't we just add a clause to opt/2?