@josevalim 2015-02-18
@nox I was able to add a clause to opt/3 as follows:

```erlang
%%% Merge label,jump,label,jump instructions to avoid multiple
%%% fixpoint calculations. Otherwise N patterns will require N*N
%%% passes. By pre-merging, we will require N passes.
opt([{label,_}=Lbl,{jump,{f,X}}|[{label,_},{jump,{f,X}}|_]=T], Acc, St) ->
    opt([Lbl|T], Acc, St);
```

I have measured the same performance benefits. My issue with it is that this is happening inside the fixpoint calculation so I am less confident about the correctness of the code.