@bjorng 2015-02-19
Your statement about correctness inspired me to look at the code I wrote more than 15 years ago in beam_jump and try to understand how it works.

The result of that deep code dive is the following alternate correction of the performance problem:

https://github.com/bjorng/otp/compare/erlang:master...bjorn/compiler/beam_jump

Please try it and verify that it solves your original performance problem.