@josevalim 2015-02-19
Yup, it solves my original performance problem too!